refactor(integrations): rename @domain/slack to @domain/integrations
Domain packages in this repo are organised by capability, not by vendor
(`@domain/email`, `@domain/marketing`, `@domain/ai` — vendors live in
their `@platform/*` adapters). `@domain/slack` was the only vendor-named
domain. Renaming it signals that future workspace integrations
(Telegram, Discord, GitHub Apps, …) will land alongside Slack here.

No behaviour or shape changes — the entity is still `SlackIntegration`,
the port is still `SlackIntegrationRepository`, etc. When a second
vendor materialises we'll have two concrete shapes to compare and can
extract a shared base then; designing the abstraction up front with one
example is premature.

The Effect service tag also moves from `@domain/slack/...` to
`@domain/integrations/...` to match.
