+++ packages/domain/integrations/src/entities/slack-integration.ts
