+++ packages/domain/integrations/src/errors.ts
