+++ packages/domain/integrations/src/use-cases/install-slack-integration.ts
