+++ packages/domain/integrations/src/use-cases/revoke-slack-integration.ts
