+++ packages/domain/integrations/tsconfig.json
