+++ packages/platform/db-postgres/package.json
@@ -47,9 +47,9 @@
     "@domain/projects": "workspace:*",
     "@domain/saved-searches": "workspace:*",
     "@domain/scores": "workspace:*",
+    "@domain/integrations": "workspace:*",
     "@domain/shared": "workspace:*",
     "@domain/simulations": "workspace:*",
-    "@domain/slack": "workspace:*",
     "@domain/spans": "workspace:*",
     "@domain/users": "workspace:*",
     "@paralleldrive/cuid2": "catalog:",
