+++ packages/platform/db-postgres/src/repositories/slack-integration-repository.test.ts
@@ -1,5 +1,5 @@
 import { generateId, OrganizationId, SlackIntegrationId, type SqlClient, UserId } from "@domain/shared"
-import { type SlackIntegration, SlackIntegrationConflictError, SlackIntegrationRepository } from "@domain/slack"
+import { type SlackIntegration, SlackIntegrationConflictError, SlackIntegrationRepository } from "@domain/integrations"
 import { Cause, Effect, Exit } from "effect"
 import { afterEach, beforeAll, describe, expect, it } from "vitest"
 import { slackIntegrations } from "../schema/slack-integrations.ts"
