+++ packages/platform/slack/scripts/slack-install.ts
@@ -2,7 +2,7 @@ import { createInterface } from "node:readline/promises"
 import { parseArgs } from "node:util"
 import { OrganizationId, UserId } from "@domain/shared"
 import { SEED_ORG_ID, SEED_OWNER_USER_ID } from "@domain/shared/seeding"
-import { installSlackIntegrationUseCase } from "@domain/slack"
+import { installSlackIntegrationUseCase } from "@domain/integrations"
 import {
   closePostgres,
   createPostgresClient,
